requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.2
PyJWT==2.6.0
